% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(x)
}
\arguments{
\item{x}{A dataset to clean.}
}
\value{
A cleaned dataset.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Removes spaces and special characters from character column(s).
}
\examples{
sos_data <- clean_data(sos_data)
}
