% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crevent.R
\name{create_crevent}
\alias{create_crevent}
\title{Create event for competing risk analysis (cmprsk::crr)}
\usage{
create_crevent(event, compevent, eventvalues = c("Yes", "Yes"))
}
\arguments{
\item{event}{Event of interest.}

\item{compevent}{Competing event (assumed to occur after event of interest).}

\item{eventvalues}{Vector with values indicating event in event and compevent (in that order).}
}
\value{
Competing event variable where 1 = event, 2 = competing event, 0 = censoring.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
suppressMessages(library(dplyr, quietly = TRUE), classes = "warnings")
rs_data <- rs_data \%>\% mutate(event_cr = create_crevent(out_hosphf, out_death_fac, eventvalues = c(1, "Yes")))
}
