% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crvar.R
\name{create_crvar}
\alias{create_crvar}
\title{Create numeric dummy xvars for competing risk analysis (cmprsk::crr)}
\usage{
create_crvar(data, var, forcenum = FALSE)
}
\arguments{
\item{data}{Data.}

\item{var}{String containing variable that should be numeric and dummified.}

\item{forcenum}{Should variables that are not character/factor be dummified? Default is FALSE.}
}
\value{
Data with a numeric xvars x levels - 1.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
rs_data <- create_crvar(rs_data, "xvar_4_num", forcenum = TRUE)
}
