% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_deathvar.R
\name{create_deathvar}
\alias{create_deathvar}
\title{Calculates death from SoS data}
\usage{
create_deathvar(cohortdata, sosdata, patid, indexdate, censdate,
  calctimetodeath = FALSE, name, orsakkod, orsakvar = ULORSAK,
  valsclass = "num", meta_reg = "Dodsorsaksregistret", meta_pos,
  warnings = TRUE)
}
\arguments{
\item{cohortdata}{Data containing the cohort.}

\item{sosdata}{Optional data where ORSAK is found. If not supplied will use cohortdata.}

\item{patid}{Patient identifier. Must be given if both cohortdata and sosdata are supplied.}

\item{indexdate}{Index date for the patient (usually date of admission or
discharge entered into SwedeHF). Must be given if calctimetodeath = TRUE.}

\item{censdate}{Vector of dates. Deaths are
allowed up until this day (usually date of death or end follow-up).}

\item{calctimetodeath}{Should time to death be calculated? Default is FALSE.}

\item{name}{Name of resulting variable
(prefix is sos_out_death).}

\item{orsakkod}{String of icd codes as a regular expression
defining the cause specific death. Should start
with blank space " ".}

\item{orsakvar}{Column where diakod is found. All codes should start
with blank space " ". Default is ULORSAK.}

\item{valsclass}{optional argument specifying the class of the
comorbidities/outcomes. Allowed values are "num" (numeric: 1, 0),
"char" (character: "yes", "no"), "fac" (factor: "yes", "no").
Default is "num".}

\item{meta_reg}{Optional argument specifying registries used. Printed in
metadatatable. Default is
"Dödsorsaksregistret".}

\item{meta_pos}{Optional argument specifying positions used to search for
outcomes/comorbidities. Printed in metadatatable.}

\item{warnings}{Should warnings be printed. Default is FALSE.}
}
\value{
dataset with column containing cause-specifc death.
Also dataset metaout that writes directly to global enviroment
with information on constructed variables.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Calculates cause-specific death (and optionally time to death)
from codes in the Cause of Death Registry.
}
\examples{

dors_data <- prep_sosdata(dors_data, registry = "dors")

rs_data <- create_deathvar(
  sosdata = dors_data,
  cohortdata = rs_data,
  patid = id,
  name = "cv",
  orsakkod = " I"
)
}
\seealso{
\code{\link{prep_sosdata}}

\code{\link{create_sosvar}}
}
