% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_medvar.R
\name{create_medvar}
\alias{create_medvar}
\title{Calculates medicines from Dispensed Drug Registry}
\usage{
create_medvar(
  atc,
  medvar = ATC,
  medname,
  cohortdata,
  meddata,
  id = "lopnr",
  valsclass = "num",
  fromdate,
  indexdate,
  metatime = "-5mo-14days"
)
}
\arguments{
\item{atc}{String of ATC as a regular expression defining the medicine. Default is ATC.}

\item{medvar}{Name of variable where ATC are found.}

\item{medname}{Name of resulting variable
(with prefix sos_ddr_).}

\item{cohortdata}{Data to add medvar to.}

\item{meddata}{Data where ATC codes are found.}

\item{id}{Vector of post identifier (meds for each post). Default is lopnr.}

\item{valsclass}{optional argument specifying the class of the
meds. Allowed values are "num" (numeric: 1, 0),
"char" (character: "Yes", "No"), "fac" (factor: "Yes", "No").
Default is "num".}

\item{fromdate}{Optional date where post with a indexdate prior to this are set to NA (e.g. prior to start of DDR + x months).}

\item{indexdate}{If fromdate is given, variable were indexdate is found.}

\item{metatime}{For what time span are the meds selected in data? Printed in metalm. Default text is "-5mo-14days".}
}
\value{
cohortdata with column containing medicine.
Also dataset metalm that writes directly to global enviroment
with information on constructed variables.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Calculates medicines from ATC from Dispensed Drug Registry.
}
\examples{

\dontrun{
lmtmp <- left_join(
  rsdata \%>\%
    select(LopNr, shf_indexdtm),
  lmsel,
  by = "LopNr"
) \%>\%
  mutate(diff = as.numeric(EDATUM - shf_indexdtm)) \%>\%
  filter(diff >= -30.5 * 5, diff <= 14) \%>\%
  select(LopNr, shf_indexdtm, EDATUM, ATC)

rsdata <- create_medvar(atc = "^C07A", medname = "bbl", cohortdata = rsdata, meddata = lmtmp, id = c("LopNr", "shf_indexdtm"))
}

}
