% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_kable.R
\name{default_kable}
\alias{default_kable}
\title{Preferred default values for creating table with kableExtra}
\usage{
default_kable(tab, font_size = NULL, scale_down = TRUE, row.names = FALSE, ...)
}
\arguments{
\item{tab}{Number that should be formatted.}

\item{font_size}{A numeric input for table font size.}

\item{scale_down}{Is useful for super wide table. It will automatically adjust the table to page width.}

\item{row.names}{Logical: whether to include row names. Default is FALSE.}

\item{...}{Additional arguments to knitr::kable.}
}
\value{
The kabletable.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
}
\examples{
tab <- tibble::tibble(a = c(2, 3, 4), b = c(5, 6, 7))
default_kable(tab)
}
