% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dors_data}
\alias{dors_data}
\title{Data from Cause of Death Registry}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{id}{Patient id. Not unique. Integer.}
\item{ULORSAK}{Underlaying cause of death}
\item{DODSDAT}{Date of death}
}
}
\usage{
dors_data
}
\description{
dors_data aims to mimic data recived from Socialstyrelsen,
used in the examples given in functions prep_sosdata and create_deathvar.
}
\keyword{datasets}
