% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\title{Format numbers (including p-values)}
\usage{
fn(x, dig = 0, p = FALSE, pequalsign = FALSE, ...)
}
\arguments{
\item{x}{Number that should be formatted.}

\item{dig}{Number of digits.}

\item{p}{is x a p-value? If <0.x1 will then format to <0.x1.}

\item{pequalsign}{if x is a p-value should = be placed in front if not <0.x1}

\item{...}{Additional arguments to formatC.}
}
\value{
The formatted number.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
x <- 1.234
fn(x)
}
