% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfmisc.R
\docType{package}
\name{hfmisc}
\alias{hfmisc}
\alias{hfmisc-package}
\title{hfmisc: Miscellaneous functions created mainly for use in RiksSvikt (SwedeHF)}
\description{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/linabe/hfmisc}
  \item \url{https://linabe.github.io/hfmisc}
  \item Report bugs at \url{https://github.com/linabe/hfmisc/issues}
}

}
\author{
Lina Benson \email{lina.benson@ki.se}
}
