% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med_data}
\alias{med_data}
\title{Data from Dispensed Drug Registry}
\format{
A data frame with 500 rows and 10 variables:
\describe{
\item{id}{Patient id. Not unique. Integer.}
\item{ATC}{ATC code. Character.}
\item{indexdtm}{Index date. Combination id and date is unique. Date.}
}
}
\usage{
med_data
}
\description{
med_data aims to mimic data from Dispensed Drug Registry (Läkmedelsregistret).
}
\keyword{datasets}
