% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sasdata.R
\name{read_sasdata}
\alias{read_sasdata}
\title{Read sas data}
\usage{
read_sasdata(
  path,
  filename,
  years,
  checkdups = FALSE,
  id,
  clean = TRUE,
  removeformats = TRUE,
  removelabels = TRUE
)
}
\arguments{
\item{path}{The path to the dataset.}

\item{filename}{The filename (without extension).}

\item{years}{Optional. If dataset from different yeas (for example from LISA)
bind together.}

\item{checkdups}{Should check for duplicates? Default is FALSE.}

\item{id}{If duplicates should be checked, identifier that should be unique.
Should be character ("LopNr")}

\item{clean}{Should the dataset be cleaned from spaces and special characters
in character columns? Default is TRUE.}

\item{removeformats}{Should sas formats be removed? Default is TRUE.}

\item{removelabels}{Should sas labels be removed? Default is TRUE.}
}
\value{
A dataset.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("maturing")}

Read sas data, clean and remove labels and formats
}
\examples{
\dontrun{
antalbarn <- read_sasdata(path = scbpath, filename = "lb_lev_antal_barn", years = 1999:2018, clean = FALSE, checkdups = TRUE, id = "LopNr")
}
}
