% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_text.R
\name{sanitize_text}
\alias{sanitize_text}
\title{Sanitize text}
\usage{
sanitize_text(str)
}
\arguments{
\item{str}{Character vector to be sanitized.}
}
\value{
The sanitized character vector.
}
\description{
Sanitize text for compatability with latex. Used for example when
escape = FALSE needed in kableExtra table (with output latex),
for example when needing footnotes.
}
\examples{
var_names <- c(">50", "#Test", "Test1 & Test2")
var_names <- sanitize_text(var_names)

}
