% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sos_data}
\alias{sos_data}
\title{Data from National Patient Registry}
\format{
A data frame with 5500 rows and 8 variables:
\describe{
\item{id}{Patient id. Not unique. Integer.}
\item{INDATUM}{Date of visit. Date.}
\item{UTDATUM}{Date of discharge. Date.}
\item{HDIA}{Main diagnosis. Character.}
\item{DIA01}{Character.}
\item{OP01}{Procedure code, position 1. Character.}
\item{OP02}{Procedure code, position 2. Character.}
\item{ekod1}{Character.}
\item{ekod2}{Character.}
\item{AR}{Year of visit date. Numeric.}
}
}
\usage{
sos_data
}
\description{
sos_data aims to mimic data recived from Socialstyrelsen,
used in the examples given in functions prep_sosdata and create_sosvar.
}
\keyword{datasets}
