% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data}
\usage{
clean_data(x)
}
\arguments{
\item{x}{A dataset to clean.}
}
\value{
A cleaned dataset.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Removes spaces and special characters from character column(s).
}
\examples{
sos_data <- clean_data(sos_data)
}
