% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\title{Format numbers (including p-values)}
\usage{
fn(x, dig = 0, p = FALSE, ...)
}
\arguments{
\item{x}{Number that should be formatted.}

\item{dig}{Number of digits.}

\item{p}{is x a p-value? If 0.000 will then format to <0.001.}

\item{...}{Additional arguments to formatC.}
}
\value{
The formatted number.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
}
\examples{
x <- 1.234
fn(x)
}
