% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sosdata.R
\name{prep_sosdata}
\alias{prep_sosdata}
\title{Prepares SoS data for input into function create_sosvar}
\usage{
prep_sosdata(sosdata, registry = "patreg", utdatum = FALSE,
  remove = TRUE, impute = TRUE)
}
\arguments{
\item{sosdata}{Data from Socialstyrelsen that should be prepared.
Should contain at least HDIA, OP, ekod, INDATUM}

\item{registry}{What registry is being prepared?
Takes values patreg (National patient registry) or dors
(Cause of death registry). Default is National patient registry.}

\item{utdatum}{Should the sosdtm (usually to be used in comparison with indexdate)
be date of discharge for hospitalisation? Default is FALSE.
Else date of admission is used.}

\item{remove}{Should the individual DIA (except HDIA), OP, ekod variables
be removed after constructing combined variable (to save space)?
Default is TRUE.}

\item{impute}{To be implemented. kom ihåg att kolla om finns värden i INDATUMA
(text) men inte i INDATUM (numerisk). i så fall imputera.}
}
\value{
sosdataset with combined DIA, OP, ekod variables and space
in front of HDIA and sosdtm.
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Prepares raw data recieved from Socialstyrelsen (national patient registry or
cause of death registry) to be processed by function make_sosvar. Removes
post with missing information on relevant dates and joins diagnosis columns
together. Reason for preparation is to save space in
internal memory (by joining columns together) and removing surplus columns.
usethis::use_lifecycle_badge("stable")
}
\examples{
\donttest{
sos_data <- bind_rows(
  sv \%>\% mutate(source = "SV"),
  ov \%>\% mutate(source = "OV"),
  dagkir \%>\% mutate(source = "DK")
)
}
sos_data <- prep_sosdata(sos_data)
}
\seealso{
\code{\link{create_sosvar}}
}
